simple: use simple sample only and adapt topology

+++ Assets/Votyra/Core/TerrainMeshes/TerrainMeshExtensions.cs
@@ -34,7 +34,7 @@ namespace Votyra.Core.TerrainMeshes
             }
         }
 
-        public static void AddQuad(this ITerrainMesh mesh, Vector2i position, SampledData2h data, SampledMask2e maskData)
+        public static void AddQuad(this ITerrainMesh mesh, Vector2f position, SampledData2h data, SampledMask2e maskData)
         {
             var x0y0 = maskData.x0y0.IsNotHole() ? new Vector2f(position.X, position.Y).ToVector3f(data.x0y0) : (Vector3f?)null;
             var x0y1 = maskData.x0y1.IsNotHole() ? new Vector2f(position.X, position.Y + 1).ToVector3f(data.x0y1) : (Vector3f?)null;
